@@ -1,19 +1,22 @@
 module vistara-node
 
-go 1.21.5
+go 1.22.0
+
+toolchain go1.22.4
 
 require (
 	github.com/containerd/typeurl/v2 v2.1.1
 	github.com/coreos/go-iptables v0.7.0
 	github.com/diskfs/go-diskfs v1.4.0
+	github.com/docker/docker v27.0.0+incompatible
 	github.com/docker/go-units v0.5.0
 	github.com/firecracker-microvm/firecracker-go-sdk v1.0.0
 	github.com/google/go-cmp v0.6.0
 	github.com/google/uuid v1.6.0
 	github.com/lorenzosaino/go-sysctl v0.3.1
 	github.com/opencontainers/go-digest v1.0.0
-	github.com/opencontainers/image-spec v1.1.0-rc3
-	github.com/prometheus/client_golang v1.14.0
+	github.com/opencontainers/image-spec v1.1.0
+	github.com/prometheus/client_golang v1.19.1
 	github.com/spf13/viper v1.19.0
 	google.golang.org/grpc v1.64.0
 	google.golang.org/protobuf v1.34.1
