@@ -3,7 +3,7 @@ package shared
 import (
 	"bytes"
 	"fmt"
-	"io/ioutil"
+	"io"
 	"os"
 	"strconv"
 	"vistara-node/pkg/defaults"
@@ -17,7 +17,7 @@ func PIDReadFromFile(pidFile string, fs afero.Fs) (int, error) {
 		return -1, fmt.Errorf("opening pid file %s: %w", pidFile, err)
 	}
 
-	data, err := ioutil.ReadAll(file)
+	data, err := io.ReadAll(file)
 	if err != nil {
 		return -1, fmt.Errorf("reading pid file %s: %w", pidFile, err)
 	}
