@@ -13,15 +13,25 @@ func convertMicroVMToModel(spec *types.MicroVMSpec) (*models.MicroVM, error) {
 		return nil, fmt.Errorf("creating vmid from spec: %w", err)
 	}
 
+	strOrDefault := func(s *string) string {
+		if s != nil {
+			return *s
+		}
+
+		return ""
+	}
+
 	convertedModel := &models.MicroVM{
 		ID: *vmid,
 		Spec: models.MicroVMSpec{
-			Kernel:     spec.KernelPath,
+			Kernel:     strOrDefault(spec.KernelPath),
 			VCPU:       spec.Vcpu,
 			MemoryInMb: spec.MemoryInMb,
-			HostNetDev: spec.HostNetDev,
-			RootfsPath: spec.RootfsPath,
-			GuestMAC:   spec.GuestMac,
+			HostNetDev: strOrDefault(spec.HostNetDev),
+			RootfsPath: strOrDefault(spec.RootfsPath),
+			GuestMAC:   strOrDefault(spec.GuestMac),
+			ImageRef:   strOrDefault(spec.ImageRef),
+			Provider:   spec.Provider,
 		},
 	}
 
@@ -41,15 +51,12 @@ func convertModelToMicroVMSpec(mvm *models.MicroVM) *types.MicroVMSpec {
 		Id:         mvm.ID.Name(),
 		Vcpu:       mvm.Spec.VCPU,
 		MemoryInMb: mvm.Spec.MemoryInMb,
-		KernelPath: mvm.Spec.Kernel,
-		RootfsPath: mvm.Spec.RootfsPath,
-		HostNetDev: mvm.Spec.HostNetDev,
-		GuestMac:   mvm.Spec.GuestMAC,
+		KernelPath: &mvm.Spec.Kernel,
+		RootfsPath: &mvm.Spec.RootfsPath,
+		HostNetDev: &mvm.Spec.HostNetDev,
+		GuestMac:   &mvm.Spec.GuestMAC,
+		ImageRef:   &mvm.Spec.ImageRef,
 	}
 
 	return converted
-}
-
-func String(s string) *string {
-	return &s
 }
