@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
 // 	protoc-gen-go v1.30.0
-// 	protoc        v4.25.3
+// 	protoc        v5.27.1
 // source: pkg/api/services/vms.proto
 
 package microvm
