@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
 // - protoc-gen-go-grpc v1.2.0
-// - protoc             v4.25.3
+// - protoc             v5.27.1
 // source: pkg/api/services/vms.proto
 
 package microvm
