@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
 // 	protoc-gen-go v1.30.0
-// 	protoc        v4.25.3
+// 	protoc        v5.27.1
 // source: pkg/api/types/microvm.proto
 
 package types
@@ -140,22 +140,23 @@ type MicroVMSpec struct {
 	// MemoryInMb is the amount of memory in megabytes that the machine will be allocated.
 	MemoryInMb int32 `protobuf:"varint,3,opt,name=memory_in_mb,json=memoryInMb,proto3" json:"memory_in_mb,omitempty"`
 	// Kernel is the details of the kernel to use .
-	KernelPath string `protobuf:"bytes,4,opt,name=kernel_path,json=kernelPath,proto3" json:"kernel_path,omitempty"`
+	KernelPath *string `protobuf:"bytes,4,opt,name=kernel_path,json=kernelPath,proto3,oneof" json:"kernel_path,omitempty"`
 	// RootVolume specifies the root volume mount for the MicroVM.
-	RootfsPath string `protobuf:"bytes,5,opt,name=rootfs_path,json=rootfsPath,proto3" json:"rootfs_path,omitempty"`
+	RootfsPath *string `protobuf:"bytes,5,opt,name=rootfs_path,json=rootfsPath,proto3,oneof" json:"rootfs_path,omitempty"`
 	// HostNetDev is the device to use for passing traffic through the TAP device
-	HostNetDev string `protobuf:"bytes,6,opt,name=HostNetDev,proto3" json:"HostNetDev,omitempty"`
+	HostNetDev *string `protobuf:"bytes,6,opt,name=HostNetDev,proto3,oneof" json:"HostNetDev,omitempty"`
 	// MAC address of the guest interface
-	GuestMac string `protobuf:"bytes,7,opt,name=GuestMac,proto3" json:"GuestMac,omitempty"`
+	GuestMac *string `protobuf:"bytes,7,opt,name=GuestMac,proto3,oneof" json:"GuestMac,omitempty"`
+	// Reference of the container to spawn
+	ImageRef *string `protobuf:"bytes,8,opt,name=ImageRef,proto3,oneof" json:"ImageRef,omitempty"`
+	// Provider allows you to specify the name of the microvm provider to use
+	Provider string `protobuf:"bytes,9,opt,name=provider,proto3" json:"provider,omitempty"`
 	// CreatedAt indicates the time the microvm was created at.
-	CreatedAt *timestamppb.Timestamp `protobuf:"bytes,12,opt,name=created_at,json=createdAt,proto3" json:"created_at,omitempty"`
+	CreatedAt *timestamppb.Timestamp `protobuf:"bytes,10,opt,name=created_at,json=createdAt,proto3" json:"created_at,omitempty"`
 	// UpdatedAt indicates the time the microvm was last updated.
-	UpdatedAt *timestamppb.Timestamp `protobuf:"bytes,13,opt,name=updated_at,json=updatedAt,proto3" json:"updated_at,omitempty"`
+	UpdatedAt *timestamppb.Timestamp `protobuf:"bytes,11,opt,name=updated_at,json=updatedAt,proto3" json:"updated_at,omitempty"`
 	// DeletedAt indicates the time the microvm was marked as deleted.
-	DeletedAt *timestamppb.Timestamp `protobuf:"bytes,14,opt,name=deleted_at,json=deletedAt,proto3" json:"deleted_at,omitempty"`
-	// Provider allows you to specify the name of the microvm provider to use. If this isn't supplied
-	// then the default provider will be used.
-	Provider *string `protobuf:"bytes,16,opt,name=provider,proto3,oneof" json:"provider,omitempty"`
+	DeletedAt *timestamppb.Timestamp `protobuf:"bytes,12,opt,name=deleted_at,json=deletedAt,proto3" json:"deleted_at,omitempty"`
 }
 
 func (x *MicroVMSpec) Reset() {
@@ -212,29 +213,43 @@ func (x *MicroVMSpec) GetMemoryInMb() int32 {
 }
 
 func (x *MicroVMSpec) GetKernelPath() string {
-	if x != nil {
-		return x.KernelPath
+	if x != nil && x.KernelPath != nil {
+		return *x.KernelPath
 	}
 	return ""
 }
 
 func (x *MicroVMSpec) GetRootfsPath() string {
-	if x != nil {
-		return x.RootfsPath
+	if x != nil && x.RootfsPath != nil {
+		return *x.RootfsPath
 	}
 	return ""
 }
 
 func (x *MicroVMSpec) GetHostNetDev() string {
-	if x != nil {
-		return x.HostNetDev
+	if x != nil && x.HostNetDev != nil {
+		return *x.HostNetDev
 	}
 	return ""
 }
 
 func (x *MicroVMSpec) GetGuestMac() string {
+	if x != nil && x.GuestMac != nil {
+		return *x.GuestMac
+	}
+	return ""
+}
+
+func (x *MicroVMSpec) GetImageRef() string {
+	if x != nil && x.ImageRef != nil {
+		return *x.ImageRef
+	}
+	return ""
+}
+
+func (x *MicroVMSpec) GetProvider() string {
 	if x != nil {
-		return x.GuestMac
+		return x.Provider
 	}
 	return ""
 }
@@ -260,13 +275,6 @@ func (x *MicroVMSpec) GetDeletedAt() *timestamppb.Timestamp {
 	return nil
 }
 
-func (x *MicroVMSpec) GetProvider() string {
-	if x != nil && x.Provider != nil {
-		return *x.Provider
-	}
-	return ""
-}
-
 var File_pkg_api_types_microvm_proto protoreflect.FileDescriptor
 
 var file_pkg_api_types_microvm_proto_rawDesc = []byte{
@@ -282,36 +290,43 @@ var file_pkg_api_types_microvm_proto_rawDesc = []byte{
 	0x52, 0x75, 0x6e, 0x74, 0x69, 0x6d, 0x65, 0x44, 0x61, 0x74, 0x61, 0x12, 0x2b, 0x0a, 0x11, 0x6e,
 	0x65, 0x74, 0x77, 0x6f, 0x72, 0x6b, 0x5f, 0x69, 0x6e, 0x74, 0x65, 0x72, 0x66, 0x61, 0x63, 0x65,
 	0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x10, 0x6e, 0x65, 0x74, 0x77, 0x6f, 0x72, 0x6b, 0x49,
-	0x6e, 0x74, 0x65, 0x72, 0x66, 0x61, 0x63, 0x65, 0x22, 0xb0, 0x03, 0x0a, 0x0b, 0x4d, 0x69, 0x63,
+	0x6e, 0x74, 0x65, 0x72, 0x66, 0x61, 0x63, 0x65, 0x22, 0x9c, 0x04, 0x0a, 0x0b, 0x4d, 0x69, 0x63,
 	0x72, 0x6f, 0x56, 0x4d, 0x53, 0x70, 0x65, 0x63, 0x12, 0x0e, 0x0a, 0x02, 0x69, 0x64, 0x18, 0x01,
 	0x20, 0x01, 0x28, 0x09, 0x52, 0x02, 0x69, 0x64, 0x12, 0x12, 0x0a, 0x04, 0x76, 0x63, 0x70, 0x75,
 	0x18, 0x02, 0x20, 0x01, 0x28, 0x05, 0x52, 0x04, 0x76, 0x63, 0x70, 0x75, 0x12, 0x20, 0x0a, 0x0c,
 	0x6d, 0x65, 0x6d, 0x6f, 0x72, 0x79, 0x5f, 0x69, 0x6e, 0x5f, 0x6d, 0x62, 0x18, 0x03, 0x20, 0x01,
-	0x28, 0x05, 0x52, 0x0a, 0x6d, 0x65, 0x6d, 0x6f, 0x72, 0x79, 0x49, 0x6e, 0x4d, 0x62, 0x12, 0x1f,
+	0x28, 0x05, 0x52, 0x0a, 0x6d, 0x65, 0x6d, 0x6f, 0x72, 0x79, 0x49, 0x6e, 0x4d, 0x62, 0x12, 0x24,
 	0x0a, 0x0b, 0x6b, 0x65, 0x72, 0x6e, 0x65, 0x6c, 0x5f, 0x70, 0x61, 0x74, 0x68, 0x18, 0x04, 0x20,
-	0x01, 0x28, 0x09, 0x52, 0x0a, 0x6b, 0x65, 0x72, 0x6e, 0x65, 0x6c, 0x50, 0x61, 0x74, 0x68, 0x12,
-	0x1f, 0x0a, 0x0b, 0x72, 0x6f, 0x6f, 0x74, 0x66, 0x73, 0x5f, 0x70, 0x61, 0x74, 0x68, 0x18, 0x05,
-	0x20, 0x01, 0x28, 0x09, 0x52, 0x0a, 0x72, 0x6f, 0x6f, 0x74, 0x66, 0x73, 0x50, 0x61, 0x74, 0x68,
-	0x12, 0x1e, 0x0a, 0x0a, 0x48, 0x6f, 0x73, 0x74, 0x4e, 0x65, 0x74, 0x44, 0x65, 0x76, 0x18, 0x06,
-	0x20, 0x01, 0x28, 0x09, 0x52, 0x0a, 0x48, 0x6f, 0x73, 0x74, 0x4e, 0x65, 0x74, 0x44, 0x65, 0x76,
-	0x12, 0x1a, 0x0a, 0x08, 0x47, 0x75, 0x65, 0x73, 0x74, 0x4d, 0x61, 0x63, 0x18, 0x07, 0x20, 0x01,
-	0x28, 0x09, 0x52, 0x08, 0x47, 0x75, 0x65, 0x73, 0x74, 0x4d, 0x61, 0x63, 0x12, 0x39, 0x0a, 0x0a,
-	0x63, 0x72, 0x65, 0x61, 0x74, 0x65, 0x64, 0x5f, 0x61, 0x74, 0x18, 0x0c, 0x20, 0x01, 0x28, 0x0b,
-	0x32, 0x1a, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62,
-	0x75, 0x66, 0x2e, 0x54, 0x69, 0x6d, 0x65, 0x73, 0x74, 0x61, 0x6d, 0x70, 0x52, 0x09, 0x63, 0x72,
-	0x65, 0x61, 0x74, 0x65, 0x64, 0x41, 0x74, 0x12, 0x39, 0x0a, 0x0a, 0x75, 0x70, 0x64, 0x61, 0x74,
-	0x65, 0x64, 0x5f, 0x61, 0x74, 0x18, 0x0d, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1a, 0x2e, 0x67, 0x6f,
-	0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x54, 0x69,
-	0x6d, 0x65, 0x73, 0x74, 0x61, 0x6d, 0x70, 0x52, 0x09, 0x75, 0x70, 0x64, 0x61, 0x74, 0x65, 0x64,
-	0x41, 0x74, 0x12, 0x39, 0x0a, 0x0a, 0x64, 0x65, 0x6c, 0x65, 0x74, 0x65, 0x64, 0x5f, 0x61, 0x74,
-	0x18, 0x0e, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1a, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e,
-	0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x54, 0x69, 0x6d, 0x65, 0x73, 0x74, 0x61,
-	0x6d, 0x70, 0x52, 0x09, 0x64, 0x65, 0x6c, 0x65, 0x74, 0x65, 0x64, 0x41, 0x74, 0x12, 0x1f, 0x0a,
-	0x08, 0x70, 0x72, 0x6f, 0x76, 0x69, 0x64, 0x65, 0x72, 0x18, 0x10, 0x20, 0x01, 0x28, 0x09, 0x48,
-	0x00, 0x52, 0x08, 0x70, 0x72, 0x6f, 0x76, 0x69, 0x64, 0x65, 0x72, 0x88, 0x01, 0x01, 0x42, 0x0b,
-	0x0a, 0x09, 0x5f, 0x70, 0x72, 0x6f, 0x76, 0x69, 0x64, 0x65, 0x72, 0x42, 0x15, 0x5a, 0x13, 0x70,
-	0x6b, 0x67, 0x2f, 0x61, 0x70, 0x69, 0x2f, 0x74, 0x79, 0x70, 0x65, 0x73, 0x3b, 0x74, 0x79, 0x70,
-	0x65, 0x73, 0x62, 0x06, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
+	0x01, 0x28, 0x09, 0x48, 0x00, 0x52, 0x0a, 0x6b, 0x65, 0x72, 0x6e, 0x65, 0x6c, 0x50, 0x61, 0x74,
+	0x68, 0x88, 0x01, 0x01, 0x12, 0x24, 0x0a, 0x0b, 0x72, 0x6f, 0x6f, 0x74, 0x66, 0x73, 0x5f, 0x70,
+	0x61, 0x74, 0x68, 0x18, 0x05, 0x20, 0x01, 0x28, 0x09, 0x48, 0x01, 0x52, 0x0a, 0x72, 0x6f, 0x6f,
+	0x74, 0x66, 0x73, 0x50, 0x61, 0x74, 0x68, 0x88, 0x01, 0x01, 0x12, 0x23, 0x0a, 0x0a, 0x48, 0x6f,
+	0x73, 0x74, 0x4e, 0x65, 0x74, 0x44, 0x65, 0x76, 0x18, 0x06, 0x20, 0x01, 0x28, 0x09, 0x48, 0x02,
+	0x52, 0x0a, 0x48, 0x6f, 0x73, 0x74, 0x4e, 0x65, 0x74, 0x44, 0x65, 0x76, 0x88, 0x01, 0x01, 0x12,
+	0x1f, 0x0a, 0x08, 0x47, 0x75, 0x65, 0x73, 0x74, 0x4d, 0x61, 0x63, 0x18, 0x07, 0x20, 0x01, 0x28,
+	0x09, 0x48, 0x03, 0x52, 0x08, 0x47, 0x75, 0x65, 0x73, 0x74, 0x4d, 0x61, 0x63, 0x88, 0x01, 0x01,
+	0x12, 0x1f, 0x0a, 0x08, 0x49, 0x6d, 0x61, 0x67, 0x65, 0x52, 0x65, 0x66, 0x18, 0x08, 0x20, 0x01,
+	0x28, 0x09, 0x48, 0x04, 0x52, 0x08, 0x49, 0x6d, 0x61, 0x67, 0x65, 0x52, 0x65, 0x66, 0x88, 0x01,
+	0x01, 0x12, 0x1a, 0x0a, 0x08, 0x70, 0x72, 0x6f, 0x76, 0x69, 0x64, 0x65, 0x72, 0x18, 0x09, 0x20,
+	0x01, 0x28, 0x09, 0x52, 0x08, 0x70, 0x72, 0x6f, 0x76, 0x69, 0x64, 0x65, 0x72, 0x12, 0x39, 0x0a,
+	0x0a, 0x63, 0x72, 0x65, 0x61, 0x74, 0x65, 0x64, 0x5f, 0x61, 0x74, 0x18, 0x0a, 0x20, 0x01, 0x28,
+	0x0b, 0x32, 0x1a, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f,
+	0x62, 0x75, 0x66, 0x2e, 0x54, 0x69, 0x6d, 0x65, 0x73, 0x74, 0x61, 0x6d, 0x70, 0x52, 0x09, 0x63,
+	0x72, 0x65, 0x61, 0x74, 0x65, 0x64, 0x41, 0x74, 0x12, 0x39, 0x0a, 0x0a, 0x75, 0x70, 0x64, 0x61,
+	0x74, 0x65, 0x64, 0x5f, 0x61, 0x74, 0x18, 0x0b, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1a, 0x2e, 0x67,
+	0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x54,
+	0x69, 0x6d, 0x65, 0x73, 0x74, 0x61, 0x6d, 0x70, 0x52, 0x09, 0x75, 0x70, 0x64, 0x61, 0x74, 0x65,
+	0x64, 0x41, 0x74, 0x12, 0x39, 0x0a, 0x0a, 0x64, 0x65, 0x6c, 0x65, 0x74, 0x65, 0x64, 0x5f, 0x61,
+	0x74, 0x18, 0x0c, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1a, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65,
+	0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x54, 0x69, 0x6d, 0x65, 0x73, 0x74,
+	0x61, 0x6d, 0x70, 0x52, 0x09, 0x64, 0x65, 0x6c, 0x65, 0x74, 0x65, 0x64, 0x41, 0x74, 0x42, 0x0e,
+	0x0a, 0x0c, 0x5f, 0x6b, 0x65, 0x72, 0x6e, 0x65, 0x6c, 0x5f, 0x70, 0x61, 0x74, 0x68, 0x42, 0x0e,
+	0x0a, 0x0c, 0x5f, 0x72, 0x6f, 0x6f, 0x74, 0x66, 0x73, 0x5f, 0x70, 0x61, 0x74, 0x68, 0x42, 0x0d,
+	0x0a, 0x0b, 0x5f, 0x48, 0x6f, 0x73, 0x74, 0x4e, 0x65, 0x74, 0x44, 0x65, 0x76, 0x42, 0x0b, 0x0a,
+	0x09, 0x5f, 0x47, 0x75, 0x65, 0x73, 0x74, 0x4d, 0x61, 0x63, 0x42, 0x0b, 0x0a, 0x09, 0x5f, 0x49,
+	0x6d, 0x61, 0x67, 0x65, 0x52, 0x65, 0x66, 0x42, 0x15, 0x5a, 0x13, 0x70, 0x6b, 0x67, 0x2f, 0x61,
+	0x70, 0x69, 0x2f, 0x74, 0x79, 0x70, 0x65, 0x73, 0x3b, 0x74, 0x79, 0x70, 0x65, 0x73, 0x62, 0x06,
+	0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
 }
 
 var (
