@@ -4,9 +4,7 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"vistara-node/pkg/api/events"
 	"vistara-node/pkg/api/types"
-	"vistara-node/pkg/defaults"
 	"vistara-node/pkg/log"
 	"vistara-node/pkg/models"
 	"vistara-node/pkg/ports"
@@ -42,13 +40,8 @@ func (a *App) Create(ctx context.Context, vm *models.MicroVM) (*models.MicroVM, 
 		return nil, errors.New("empty vmID")
 	}
 
-	if vm.Spec.Provider == "" {
-		vm.Spec.Provider = a.cfg.DefaultProvider
-	}
-
 	logger.Infof("Hypervisor provider: %s", vm.Spec.Provider)
 
-	// vm.Spec.CreatedAt = a.ports.Clock().Unix()
 	vm.Status.State = models.PendingState
 	vm.Status.Retry = 0
 
@@ -58,16 +51,14 @@ func (a *App) Create(ctx context.Context, vm *models.MicroVM) (*models.MicroVM, 
 		return nil, fmt.Errorf("saving microvm: %w", err)
 	}
 
-	// Publish a MicroVMCreated event to a queue.
-	if err := a.ports.EventService.Publish(
-		ctx,
-		defaults.TopicMicroVMEvents,
-		&events.MicroVMSpecCreated{
-			ID:        vm.ID.Name(),
-			Namespace: vm.ID.Namespace(),
-			UID:       vm.ID.UID(),
-		}); err != nil {
-		return nil, fmt.Errorf("publishing microVM created event, vmid %s: %w", vm.ID.Name(), err)
+	provider, ok := a.ports.MicrovmProviders[vm.Spec.Provider]
+	if !ok {
+		return nil, fmt.Errorf("provider %s not found", vm.Spec.Provider)
+	}
+
+	// call chosen hypervisor to start the vm
+	if err := provider.Start(ctx, vm); err != nil {
+		return nil, fmt.Errorf("starting microvm: %w", err)
 	}
 
 	return createdVm, nil
@@ -93,7 +84,7 @@ func (a *App) Delete(ctx context.Context, vmid models.VMID) error {
 		UID:       vmid.UID(),
 	})
 	if err != nil {
-		return fmt.Errorf("Getting MicroVM specs: %w", err)
+		return fmt.Errorf("getting MicroVM specs: %w", err)
 	}
 
 	logger.Infof("hypervisor provider is %v", vm.Spec.Provider)
@@ -117,23 +108,6 @@ func (a *App) Reconcile(ctx context.Context, vmid models.VMID) error {
 	logger := log.GetLogger(ctx).WithField("action", "reconcile")
 
 	logger.Debugf("Creating microvm in reconcile %v\n", vmid)
-
-	vm, err := a.ports.Repo.Get(ctx, ports.RepositoryGetOptions{
-		Name:      vmid.Name(),
-		Namespace: vmid.Namespace(),
-		UID:       vmid.UID(),
-	})
-
-	if err != nil {
-		return fmt.Errorf("Getting MicroVM spec to start VM: %w", err)
-	}
-
-	logger.Infof("hypervisor provider is %v", vm.Spec.Provider)
-
-	// call chosen hypervisor to start the vm
-	if err := a.ports.MicrovmProviders[vm.Spec.Provider].Start(ctx, vm); err != nil {
-		return fmt.Errorf("starting microvm: %w", err)
-	}
 
 	return nil
 }
