@@ -1,9 +1,5 @@
 package containerd
 
-const (
-	supportedSnapshotters = "overlayfs,native,devmapper"
-)
-
 // Config holds the containerd configuration.
 type Config struct {
 	// SnapshotterKernel is the name of the containerd snapshotter to use for kernel images.
@@ -13,5 +9,6 @@ type Config struct {
 	// SocketPath is the path to the containerd socket.
 	SocketPath string
 	// Namespace is the default containerd namespace to use
-	Namespace string
+	Namespace          string
+	ContainerNamespace string
 }
