@@ -31,7 +31,7 @@ const (
 	FirecrackerDetach = true
 
 	// CloudHypervisorBin is the name of the Cloud Hypervisor binary.
-	CloudHypervisorBin = "cloud-hypervisor-static"
+	CloudHypervisorBin = "cloud-hypervisor"
 
 	// CloudHypervisorDetach is the default for the flag to indicates with the child cloud-hypervisor
 	// processes should be run detached.
