@@ -25,16 +25,6 @@ var (
 	ErrUnableToBoot                       = errors.New("microvm is unable to boot")
 )
 
-// TopicNotFoundError is an error created when a topic with a specific name isn't found.
-type TopicNotFoundError struct {
-	Name string
-}
-
-// Error returns the error message.
-func (e TopicNotFoundError) Error() string {
-	return fmt.Sprintf("topic %s not found", e.Name)
-}
-
 type IncorrectVMIDFormatError struct {
 	ActualID string
 }
@@ -44,52 +34,20 @@ func (e IncorrectVMIDFormatError) Error() string {
 	return fmt.Sprintf("unexpected vmid format: %s", e.ActualID)
 }
 
-func NewErrUnsupportedInterface(ifaceType string) UnsupportedInterfaceError {
-	return UnsupportedInterfaceError{
-		ifaceType: ifaceType,
-	}
-}
-
-type UnsupportedInterfaceError struct {
-	ifaceType string
+type specNotFoundError struct {
+       name      string
+       namespace string
+       version   string
+       uid       string
 }
 
 // Error returns the error message.
-func (e UnsupportedInterfaceError) Error() string {
-	return fmt.Sprintf("network interface type %s is unsupported", e.ifaceType)
-}
+func (e specNotFoundError) Error() string {
+       if e.version == "" {
+               return fmt.Sprintf("microvm spec %s/%s/%s not found", e.namespace, e.name, e.uid)
+       }
 
-func NewVolumeNotMounted(volumeID string) VolumeNotMountedError {
-	return VolumeNotMountedError{
-		id: volumeID,
-	}
-}
-
-// VolumeNotMountedError is an error used when a volume hasn't been mounted.
-type VolumeNotMountedError struct {
-	id string
-}
-
-// Error returns the error message.
-func (e VolumeNotMountedError) Error() string {
-	return fmt.Sprintf("volume %s is not mounted", e.id)
-}
-
-func NewNetworkInterfaceStatusMissing(guestIface string) NetworkInterfaceStatusMissingError {
-	return NetworkInterfaceStatusMissingError{
-		guestIface: guestIface,
-	}
-}
-
-// NetworkInterfaceStatusMissing is an error used when a network interfaces
-// status cannot be found.
-type NetworkInterfaceStatusMissingError struct {
-	guestIface string
-}
-
-// Error returns the error message.
-func (e NetworkInterfaceStatusMissingError) Error() string {
-	return fmt.Sprintf("status for network interface %s is not found", e.guestIface)
+       return fmt.Sprintf("microvm spec %s/%s/%s not found with version %s", e.namespace, e.name, e.uid, e.version)
 }
 
 func NewSpecNotFound(name, namespace, version, uid string) error {
@@ -99,49 +57,4 @@ func NewSpecNotFound(name, namespace, version, uid string) error {
 		version:   version,
 		uid:       uid,
 	}
-}
-
-type specNotFoundError struct {
-	name      string
-	namespace string
-	version   string
-	uid       string
-}
-
-// Error returns the error message.
-func (e specNotFoundError) Error() string {
-	if e.version == "" {
-		return fmt.Sprintf("microvm spec %s/%s/%s not found", e.namespace, e.name, e.uid)
-	}
-
-	return fmt.Sprintf("microvm spec %s/%s/%s not found with version %s", e.namespace, e.name, e.uid, e.version)
-}
-
-// IsSpecNotFound tests an error to see if its a spec not found error.
-func IsSpecNotFound(err error) bool {
-	e := &specNotFoundError{}
-
-	return errors.As(err, e)
-}
-
-func NewNotSupported(featureName string) error {
-	return notSupportedError{
-		unsupported: featureName,
-	}
-}
-
-type notSupportedError struct {
-	unsupported string
-}
-
-// Error returns the error message.
-func (e notSupportedError) Error() string {
-	return fmt.Sprintf("%s is not supported", e.unsupported)
-}
-
-// IsNotSupported tests an error to see if its a not supported error.
-func IsNotSupported(err error) bool {
-	e := &notSupportedError{}
-
-	return errors.As(err, e)
 }
