@@ -1,46 +1,47 @@
 package hypervisor
 
 import (
-	"errors"
 	"fmt"
 	"vistara-node/internal/config"
+	"vistara-node/pkg/hypervisor/cloudhypervisor"
+	"vistara-node/pkg/hypervisor/docker"
 	"vistara-node/pkg/hypervisor/firecracker"
+	"vistara-node/pkg/hypervisor/runc"
 	"vistara-node/pkg/ports"
 
 	"github.com/spf13/afero"
 )
 
-var errUnknownHypervisor = "unknown hypervisor"
-
-func New(name string, cfg *config.Config, networkSvc ports.NetworkService, fs afero.Fs) (ports.MicroVMService, error) {
-	switch name {
-	case firecracker.HypervisorName:
-		return firecracker.New(firecrackerConfig(cfg), networkSvc, fs), nil
-	default:
-		return nil, nil
-	}
-
-}
-
-func firecrackerConfig(cfg *config.Config) *firecracker.Config {
-	return &firecracker.Config{
-		FirecrackerBin: cfg.FirecrackerBin,
-		RunDetached:    cfg.FirecrackerDetatch,
-		StateRoot:      fmt.Sprintf("%s/vm", cfg.StateRootDir),
-	}
-}
-
 // NewFromConfig will create instances of the vm providers based on the config.
-func NewFromConfig(cfg *config.Config, networkSvc ports.NetworkService, diskSvc ports.DiskService, fs afero.Fs) (map[string]ports.MicroVMService, error) {
+func NewFromConfig(cfg *config.Config, networkSvc ports.NetworkService, diskSvc ports.DiskService, fs afero.Fs, containerd ports.MicroVMRepository) (map[string]ports.MicroVMService, error) {
 	providers := map[string]ports.MicroVMService{}
 
 	if cfg.FirecrackerBin != "" {
-		providers[firecracker.HypervisorName] = firecracker.New(firecrackerConfig(cfg), networkSvc, fs)
+		providers[firecracker.HypervisorName] = firecracker.New(&firecracker.Config{
+			FirecrackerBin: cfg.FirecrackerBin,
+			RunDetached:    cfg.FirecrackerDetatch,
+			StateRoot:      fmt.Sprintf("%s/vm", cfg.StateRootDir),
+		}, networkSvc, fs)
 	}
 
-	if len(providers) == 0 {
-		return nil, errors.New("you must enable at least 1 microvm provider")
+	if cfg.CloudHypervisorBin != "" {
+		providers[cloudhypervisor.HypervisorName] = cloudhypervisor.New(&cloudhypervisor.Config{
+			CloudHypervisorBin: cfg.CloudHypervisorBin,
+			RunDetached:        cfg.CloudHypervisorDetatch,
+			StateRoot:          fmt.Sprintf("%s/vm", cfg.StateRootDir),
+		}, networkSvc, diskSvc, fs)
 	}
+
+	dockerProvider, err := docker.New()
+	if err != nil {
+		return nil, fmt.Errorf("failed to create docker provider: %w", err)
+	}
+
+	providers[docker.HypervisorName] = dockerProvider
+
+	providers[runc.HypervisorName] = runc.New(&runc.Config{
+		StateRoot: fmt.Sprintf("%s/vm", cfg.StateRootDir),
+	}, containerd, fs)
 
 	return providers, nil
 }
