@@ -1,10 +1,12 @@
 package models
 
-import (
-	"fmt"
-	"net"
-	"strings"
-)
+import "net"
+
+type TapDetails struct {
+	VmIp  net.IP
+	TapIp net.IP
+	Mask  net.IP
+}
 
 // NetworkInterface represents a network interface for the microvm.
 type NetworkInterface struct {
@@ -31,40 +33,11 @@ type NetworkInterface struct {
 type StaticAddress struct {
 	// Address is the static IP address (IPv4 or IPv6) to assign to this interface.
 	// Must be CIDR notation.
-	Address IPAddressCIDR `json:"address" validate:"cidr"`
+	Address string `json:"address" validate:"cidr"`
 	// Gateway is used to optionally set the default gateway for IPv4 or IPv6.
-	Gateway *IPAddressCIDR `json:"gateway,omitempty" validate:"omitempty,cidr"`
+	Gateway *string `json:"gateway,omitempty" validate:"omitempty,cidr"`
 	// Nameservers allows you to optionally specify nameservers for the interface.
 	Nameservers []string `json:"nameservers" validate:"omitempty,dive,ip"`
-}
-
-// IPAddressCIDR represents a IPv4/IPv6 address in CIDR notation.
-type IPAddressCIDR string
-
-func (i IPAddressCIDR) IsIPv4() (bool, error) {
-	ip, _, err := net.ParseCIDR(string(i))
-	if err != nil {
-		return false, fmt.Errorf("parsing %s as cidr: %w", i, err)
-	}
-
-	// 0:0:0:0:0:ffff:0101:0101 is the ipv6 representation of 1.1.1.1, the net
-	// package can parse and convert to IPv4. Based on Slack messages, we want a
-	// strict validation, they have to spcify an IPv4 CIDR block.
-	containsDot := strings.Contains(string(i), ".")
-
-	return ip.To4() != nil && containsDot, nil
-}
-
-func (i IPAddressCIDR) IP() (string, error) {
-	if _, _, err := net.ParseCIDR(string(i)); err != nil {
-		return "", fmt.Errorf("parsing %s as cidr: %w", i, err)
-	}
-
-	// We don't have to test if we can get IPv6 or IPv4 address with To4 or
-	// To16 because ParseCIDR returns with an error if it's neither of them.
-	slashIndex := strings.Index(string(i), "/")
-
-	return string(i)[:slashIndex], nil
 }
 
 type NetworkInterfaceStatus struct {
@@ -75,6 +48,8 @@ type NetworkInterfaceStatus struct {
 	Index int `json:"index"`
 	// MACAddress is the MAC address of the host interface.
 	MACAddress string `json:"mac_address"`
+	// TAP devicee details
+	TapDetails TapDetails
 }
 
 // NetworkInterfaceStatuses is a collection of network interfaces.
